+++ src/typesold.hs
@@ -1,60 +1,51 @@
 {-# LANGUAGE DeriveGeneric #-}
 
-{- |
-Module      : Types
-Description : Declaration of classes and types for hero-wars utilities
-Copyright   : (c) Stratis Christodoulou 2025
-Maintainer  : stratis.vip@gmail.com
-Stability   : experimental
-
-This module holds all types, data definition and classes
-needed for the hero-wars utilities i am developing.
-
-Original date: 5 Apr 2025
--}
-module Typesold (
-  Member (..),
-  Pet (..),
-  Hero (..),
-  Titan (..),
-  HasNameAndId (..),
-) where
+-- |
+-- Module      : Types
+-- Description : Declaration of classes and types for hero-wars utilities
+-- Copyright   : (c) Stratis Christodoulou 2025
+-- Maintainer  : stratis.vip@gmail.com
+-- Stability   : experimental
+--
+-- This module holds all types, data definition and classes
+-- needed for the hero-wars utilities i am developing.
+--
+-- Original date: 5 Apr 2025
+module Typesold
+  ( Member (..),
+    Pet (..),
+    Titan (..),
+    HasNameAndId (..),
+  )
+where
 
 import GHC.Generics (Generic)
+import Types.Assorted (ID, PHTName)
 
 -- import Types.Guild (League (..))
-
--- | alias for ids
-type ID = Int
--- ^ alias for id's
-
-type IntDate = Int
--- ^ alias for integer dates in form of YYYYMMDD
 
 type Lineup = (String, String, String, String, String)
 -- ^ alias of a five string tuple to keeps the names of heros, pets, titans
 
-type PHTName = (String, String, Int) -- (P)et(H)ero(T)itan(Name)
 -- ^ alias of Pet - Hero - Titan triple who keeps, shortname, name, id
 
 -- | Structure to keep any member of a 'Guild' data
 data Member = Member
-  { mId :: ID
-  -- ^ unique id
-  , mName :: String
-  -- ^ the name of the hero
-  , guild :: Maybe ID
-  -- ^ the 'Guild' he belongs. This is an existing 'ID' or Nothing
-  , maxHPower :: Int
-  -- ^ max power of heros documented
-  , maxTPower :: Int
-  -- ^ max power of Titans documented
+  { -- | unique id
+    mId :: ID,
+    -- | the name of the hero
+    mName :: String,
+    -- | the 'Guild' he belongs. This is an existing 'ID' or Nothing
+    guild :: Maybe ID,
+    -- | max power of heros documented
+    maxHPower :: Int,
+    -- | max power of Titans documented
+    maxTPower :: Int
   }
   deriving (Show, Generic)
 
-{- |
- Describe a polymorphic way to get id's, names and short names from various structures
--}
+-- |
+-- Describe a polymorphic way to get id's, names and short names from various structures
 class HasNameAndId a where
   getName :: a -> String
   -- ^ get the name
@@ -62,7 +53,7 @@ class HasNameAndId a where
   getShortName :: a -> String
   -- ^ get the shortname. If no short name provided, it returns the 3 first characters of name
 
-  getId :: a -> Int
+  getId :: a -> ID
   -- ^ get the unique id
 
 instance HasNameAndId Member where
@@ -83,7 +74,6 @@ instance HasNameAndId Titan where
 {- Pets  -}
 data Pet = Pet {pDtls :: PHTName} deriving (Show)
 
-data Hero = Hero {heroDtls :: PHTName} deriving (Show)
 data Titan = Titan {titanDtls :: PHTName} deriving (Show)
 
 {-