+++ CHANGELOG.md
@@ -7,4 +7,8 @@
 ## 0.1.0.1 -- 2025-04-16
 * Change the main function to give a set of choices with what to do. 
 * Move all helping functions to helpers and make tests for all of them
-* 
+* Add csv functionality
+
+## 0.1.1.2 -- 2025-04-17
+* Add a Types.Hero module
+. Add a Types.Assorted module to keep all types that are not important for a special module, like IntDate etc